requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-debug-toolbar==4.2.0
django-rest-framework==0.1.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
drf-spectacular==0.27.1
flake8==7.0.0
flower==2.0.1
humanize==4.9.0
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mccabe==0.7.0
pillow==10.2.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
rpds-py==0.17.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
